requirements.txt
numpy
scipy
pandas
tqdm
pyzmq
opencv-python-headless
tifffile
pyqt5
pyqtgraph
pillow
scikit-video
pyserial
natsort
